refactored code to split various models per functionalitiy
