+++ package.json
@@ -2,11 +2,11 @@
   "name": "bookstore",
   "version": "1.0.0",
   "description": "A NodeJS Book Store App from Zuri Training",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js",
-    "index": "nodemon index.js"
+    "start": "node src/index.js",
+    "index": "nodemon src/index.js"
   },
   "repository": {
     "type": "git",
