+++ src/controllers/bookControllers.js
@@ -1,40 +1,6 @@
-const express = require('express');
-const app = express();
-const port = 4000;
+const Book = require("../models/book");
 
-app.use(express.json())
-
-// Setup mongoose
-const mongoose = require('mongoose');
-const connectionString = "mongodb://localhost:27017/bookapp";
-
-mongoose.connect(connectionString, {
-    useFindAndModify: false,
-    useNewUrlParser: true,
-    useUnifiedTopology: true
-}, (err) => {
-    if (err) {
-        console.log(err)
-    } else {
-        console.log("Connection to database Successful...")
-    }
-});
-
-// CREATE SCHEMA
-const bookSchema = new mongoose.Schema({
-    title: String,
-    author: String,
-    description: String,
-    category: String,
-    purchaseCount: Number,
-    imageURL: String,
-    tags: Array
-});
-
-const Book = mongoose.model('Book', bookSchema);
-
-// POST request /books to create a new book.
-app.post('/books', (req, res) => {
+exports.createNewBook = (req, res) => {
     // retrieve new book details from request body
     Book.create({
         title: req.body.title,
@@ -51,12 +17,9 @@ app.post('/books', (req, res) => {
             return res.status(200).json({message: "New Book created."})
         }
     })
-    // create and save new book into database
-    // send a response to client
-})
+}
 
-// GET request /books to fetch all books
-app.get('/books', (req, res) => {
+exports.fetchBooks = (req, res) => {
     // fetch all books
     Book.find({}, (err, books) => {
         if (err) {
@@ -65,10 +28,9 @@ app.get('/books', (req, res) => {
             return res.status(200).json({message: books})
         }
     })
-    // and send response to client
-})
-// GET request /books/id to fetch a single book
-app.get("/books/:id", (req, res) => {
+}
+
+exports.fetchSingleBook =  (req, res) => {
     Book.findOne({_id: req.params.id}, (err, book) => {
         if (err) {
             return res.status(500).json({message: err});
@@ -78,9 +40,9 @@ app.get("/books/:id", (req, res) => {
             return res.status(200).json({message: book})
         }
     })
-})
-// PUT  request /books/id to update a single book
-app.put('/books/:id', (req, res) => {
+}
+
+exports.updateSingleBook = (req, res) => {
     Book.findByIdAndUpdate(req.params.id, {
         title: req.body.title,
         author: req.body.author
@@ -99,9 +61,9 @@ app.put('/books/:id', (req, res) => {
             });
         }
     })
-})
-// DELETE request /books/id to delete a single book
-app.delete("/books/:id", (req, res) => {
+}
+
+exports.deleteSingleBook = (req, res) => {
     Book.findByIdAndDelete(req.params.id, (err, book) => {
         if (err) {
             return res.status(500).json({message: err});
@@ -111,6 +73,4 @@ app.delete("/books/:id", (req, res) => {
             return res.status(200).json({message: "Book deleted successfully!"})
         }
     })
-})
-
-app.listen(port, () => { console.log(`App is listening on port ${port}`)});
+}
